Fix tag association in tables
Global tags such as "BOOT_ROM" were only working for the first iteration
of a table, and would otherwise be empty. This resulted in invalid
linker script when more than two RAM regions were defined.

Using a String to Unbounded_String association rather the String to Tag
fixes the issue. As shown in the output of the print_tags test, the
impacted tags are no longer vectors of element.

The symbol matching tests are modified to exercise this case.

Part of W109-026.